requirements.txt
lxml==4.9.3
python-docx==1.1.0
python-dotenv==1.0.0
typing_extensions==4.8.0
